Attempt to prevent loading a protocol from two places
When protocols are consolidated, they get put in the `consolidated` directory
inside the build. This directory is added to the load-path so that they are
loaded before the unconsolidated versions in the normal path. Dialyzer, when
using the `files_rec` option might complain about a module being defined in two
different places. Therefore, we filter out the unconsolidated versions from the
file list to prevent them from being analyzed twice.